EMP-406 Added cache service
