+++ package.json
@@ -101,7 +101,6 @@
     "@azure/msal-common": "^14.13.0",
     "@azure/msal-node": "^2.10.0",
     "@ministryofjustice/frontend": "^2.1.1",
-    "@types/jsonwebtoken": "^9.0.6",
     "agentkeepalive": "^4.5.0",
     "applicationinsights": "^2.9.5",
     "axios": "^1.7.1",
@@ -120,7 +119,6 @@
     "helmet": "^7.1.0",
     "http-errors": "^2.0.0",
     "joi": "^17.13.0",
-    "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "lru-cache": "^11.0.0",
     "nocache": "^4.0.0",
