+++ server/data/api/sdsApiClient.ts
@@ -1,32 +1,10 @@
-import msal from '@azure/msal-node'
-import { ClientCredentialRequest } from '@azure/msal-node/src/request/ClientCredentialRequest'
-import { Configuration } from '@azure/msal-node/src/config/Configuration'
-import jwt from 'jsonwebtoken'
 import RestClient from '../restClient'
 import config from '../../config'
-import cacheService, { SDS_ACCESS_TOKEN } from '../../services/cacheService'
-import logger from '../../../logger'
+import authProvider from '../../auth/authProvider'
+import { sdsCacheService, SDS_ACCESS_TOKEN } from '../../services/cacheService'
 
 export type SdsResponse = {
   fileURL: string
-}
-
-const getAccessToken = async (): Promise<string> => {
-  const clientConfig: Configuration = {
-    auth: {
-      clientId: process.env.CLIENT_ID,
-      authority: `${process.env.CLOUD_INSTANCE}${process.env.TENANT_ID}`,
-      clientSecret: process.env.CLIENT_SECRET,
-    },
-  }
-  const confidentialClientApplication = new msal.ConfidentialClientApplication(clientConfig)
-  const clientCredentialRequest: ClientCredentialRequest = {
-    scopes: [process.env.SDS_AUTH_SCOPE],
-    skipCache: true,
-  }
-
-  const response = await confidentialClientApplication.acquireTokenByClientCredential(clientCredentialRequest)
-  return response.accessToken
 }
 
 export default class SdsApiClient {
@@ -35,14 +13,7 @@ export default class SdsApiClient {
   }
 
   async retrieveFile(fileKey: string): Promise<SdsResponse> {
-    const accessToken = await this.getAccessTokenFromCache()
-    //
-    // try {
-    //   const decoded = jwt.verify(accessToken, <SECRET_NEEEDED>)
-    // } catch (err) {
-    //   logger.warn(`Failed to verify access token: ${err}`)
-    // }
-
+    const accessToken = await this.getAccessToken()
     return SdsApiClient.restClient('no_auth').get<SdsResponse>({
       path: '/retrieve_file',
       headers: { authorization: `Bearer ${accessToken}` },
@@ -52,12 +23,13 @@ export default class SdsApiClient {
     })
   }
 
-  private getAccessTokenFromCache = async (): Promise<string> => {
-    if (cacheService.has(SDS_ACCESS_TOKEN)) {
-      return cacheService.get(SDS_ACCESS_TOKEN)
+  private getAccessToken = async (): Promise<string> => {
+    if (sdsCacheService.has(SDS_ACCESS_TOKEN)) {
+      return sdsCacheService.get(SDS_ACCESS_TOKEN)
     }
-    const accessToken = await getAccessToken()
-    cacheService.set(SDS_ACCESS_TOKEN, accessToken)
+
+    const accessToken = await authProvider.getSDSAccessToken()
+    sdsCacheService.set(SDS_ACCESS_TOKEN, accessToken)
     return accessToken
   }
 }
